# DictionarySlim backport improvements: remove null check on insert