--- a/src/System.Private.CoreLib/shared/System/Collections/Generic/Dictionary.cs
+++ b/src/System.Private.CoreLib/shared/System/Collections/Generic/Dictionary.cs
@@ -5,7 +5,6 @@
 using System.Diagnostics;
 using System.Runtime.CompilerServices;
 using System.Runtime.Serialization;
-using System.Threading;
 
 namespace System.Collections.Generic
 {
@@ -53,6 +52,7 @@ namespace System.Collections.Generic
         private IEqualityComparer<TKey> _comparer;
         private KeyCollection _keys;
         private ValueCollection _values;
+        private static readonly Entry[] InitialEntries = new Entry[1] { new Entry() { hashCode = -1 } };
 
         // constants for serialization
         private const string VersionName = "Version"; // Do not rename (binary serialization)
@@ -79,6 +79,15 @@ namespace System.Collections.Generic
             {
                 // To start, move off default comparer for string which is randomised
                 _comparer = (IEqualityComparer<TKey>)NonRandomizedStringEqualityComparer.Default;
+            }
+
+            // Default initialization to avoid null check for every insert
+            // Reads will fail and first add cause a resize into a real array
+            if (_entries is null)
+            {
+                _buckets = HashHelpers.SizeOneIntArray;
+                _entries = InitialEntries;
+                _freeList = -1;
             }
         }
 
@@ -474,11 +483,6 @@ namespace System.Collections.Generic
                 ThrowHelper.ThrowArgumentNullException(ExceptionArgument.key);
             }
 
-            if (_buckets == null)
-            {
-                Initialize(0);
-            }
-
             Entry[] entries = _entries;
             IEqualityComparer<TKey> comparer = _comparer;
 
@@ -624,7 +628,9 @@ namespace System.Collections.Generic
             else
             {
                 int count = _count;
-                if (count == entries.Length)
+
+                // _entries.Length == 1 is dummy entries
+                if (count == entries.Length || entries.Length == 1)
                 {
                     Resize();
                     bucket = ref _buckets[hashCode % _buckets.Length];
@@ -968,7 +974,9 @@ namespace System.Collections.Generic
         {
             if (capacity < 0)
                 ThrowHelper.ThrowArgumentOutOfRangeException(ExceptionArgument.capacity);
-            int currentCapacity = _entries == null ? 0 : _entries.Length;
+
+            // _entries.Length == 1 is dummy entries
+            int currentCapacity = _entries.Length == 1 ? 0 : _entries.Length;
             if (currentCapacity >= capacity)
                 return currentCapacity;
             _version++;
@@ -1006,7 +1014,9 @@ namespace System.Collections.Generic
             int newSize = HashHelpers.GetPrime(capacity);
 
             Entry[] oldEntries = _entries;
-            int currentCapacity = oldEntries == null ? 0 : oldEntries.Length;
+
+            // _entries.Length == 1 is dummy entries
+            int currentCapacity = _entries.Length == 1 ? 0 : oldEntries.Length;
             if (newSize >= currentCapacity)
                 return;