@jkotas 2019-02-14
> it's not easy compare as expected, sometimes I get better result and other worste.

It is explainable that you are not able to measure any throughput gains from this change. The one extra null check you are removing will be very well predicted when TryInsert is called in the loop because of it is rarely taken. It means that will costs nothing, especially given the size and complexity of the method that it is part of.

On the other hand, this increases Dictionary static footprint (adds static constructor, adds a static field, adds extra or more checks to multiple other places).

Thanks for giving it a try, but this change does not look like an improvement to me.