@MarcoRossignoli 2019-02-14
> Thanks for giving it a try, but this change does not look like an improvement to me.

I agree with all your analysis but I have to show that I did my homework :relaxed:
@danmosemsft if also you agree, could you draw a straight line on third point on issue?I'll go on with other.