@stephentoub 2019-02-15
> could you draw a straight line on third point on issue?

Done