@danmosemsft 2019-03-04
Thanks for trying. BTW this was not to improve throughput but reduce footprint of empty dictionaries.