@MarcoRossignoli 2019-03-04
> Thanks for trying. BTW this was not to improve throughput but reduce footprint of empty dictionaries.

Ah ok from issue I misunderstood `Eliminate _buckets == null check on every access` anyway seems there isn't room for improve here.